robot-framework-python-tests/requirements.txt
robotframework==7.3.1
RESTinstance==1.5.2
robotframework-seleniumlibrary==6.7.1
robotframework-browser==19.6.1
robotframework-imagehorizonlibrary==1.0.0
Pillow==11.3.0
allure-robotframework==2.14.3
robotframework-pabot==4.2.0
robotframework-dashboard==1.0.2
pytest==8.4.1
pytest-html==4.1.1
requests==2.32.4
pytest-cov==6.2.1
responses==0.25.7
jsonschema==4.23.0
numpy==1.26.4  
GitHubApi==0.1.0
